Fix Battle Only mode
relates to #2122
and also some code cleanup
